src: change class BasicShapeElements GLubytes types to GLuint
This change is need for futur work

+++ tp2_H24/src/shapes.h
@@ -6,12 +6,12 @@
 class BasicShapeArrays
 {
 public:
-    BasicShapeArrays(const GLfloat* data, GLsizeiptr byteSize);    
+    BasicShapeArrays(const GLfloat* data, GLsizeiptr byteSize);
     ~BasicShapeArrays();
-    
+
     void enableAttribute(GLuint index, GLint size, GLsizei stride, GLsizeiptr offset);
     void draw(GLenum mode, GLsizei count);
-    
+
 private:
     GLuint m_vao;
     GLuint m_vbo;
@@ -23,17 +23,17 @@ class BasicShapeMultipleArrays
 public:
     BasicShapeMultipleArrays(const GLfloat* pos, GLsizeiptr posByteSize, const GLfloat* color, GLsizeiptr colorByteSize);
     ~BasicShapeMultipleArrays();
-    
-    void enablePosAttribute(GLuint index, GLint size, GLsizei stride, GLsizeiptr offset);    
+
+    void enablePosAttribute(GLuint index, GLint size, GLsizei stride, GLsizeiptr offset);
     void enableColorAttribute(GLuint index, GLint size, GLsizei stride, GLsizeiptr offset);
-    
+
     void updateColorData(const GLfloat* color, GLsizeiptr colorByteSize);
-    
-    GLfloat* mapPosData();    
+
+    GLfloat* mapPosData();
     void unmapPosData();
-    
+
     void draw(GLenum mode, GLsizei count);
-    
+
 private:
     GLuint m_vao;
     GLuint m_posVbo;
@@ -44,12 +44,12 @@ private:
 class BasicShapeElements
 {
 public:
-    BasicShapeElements(const GLfloat* data, GLsizeiptr byteSize, const GLubyte* indexes, GLsizeiptr indexesByteSize);    
+    BasicShapeElements(const GLfloat* data, GLsizeiptr byteSize, const GLuint* indexes, GLsizeiptr indexesByteSize);
     ~BasicShapeElements();
-    
-    void enableAttribute(GLuint index, GLint size, GLsizei stride, GLsizeiptr offset);    
+
+    void enableAttribute(GLuint index, GLint size, GLsizei stride, GLsizeiptr offset);
     void draw(GLenum mode, GLsizei count);
-    
+
 private:
     GLuint m_vao;
     GLuint m_vbo;
